+++ db/upgrade.php
@@ -12,6 +12,8 @@ require __DIR__ . '/../plugin/bootstrap/autoload.php';
  */
 function xmldb_charon_upgrade($oldversion = 0)
 {
+    clear_caches();
+
     global $DB;
     $dbManager = $DB->get_manager();
 
@@ -813,7 +815,7 @@ function xmldb_charon_upgrade($oldversion = 0)
         }
     }
 
-    if ($oldversion < 2021120203) {
+    if ($oldversion < 2022021001) {
         $table = new xmldb_table("charon");
         $field = new xmldb_field('unittests_git', XMLDB_TYPE_CHAR, 255, null, null, null, null, null, null);
 
@@ -822,5 +824,36 @@ function xmldb_charon_upgrade($oldversion = 0)
         }
     }
 
+    if ($oldversion < 2022031101) {
+        $sql = "CREATE TABLE " . $CFG->prefix . "charon_query_log_user(" .
+            "    id BIGINT(10) AUTO_INCREMENT NOT NULL," .
+            "    user_id BIGINT(10) NOT NULL," .
+            "    PRIMARY KEY (id)," .
+            "    CONSTRAINT FK_query_log_user_user" .
+            "        FOREIGN KEY (user_id)" .
+            "            REFERENCES " . $CFG->prefix . "user(id)" .
+            "            ON DELETE CASCADE" .
+            "            ON UPDATE CASCADE" .
+            ")";
+
+        $table = new xmldb_table("charon_query_log_user");
+
+        if (!$dbManager->table_exists($table)) {
+            $DB->execute($sql);
+        }
+    }
+
     return true;
+}
+
+function clear_caches() {
+    try {
+        require __DIR__ . '/../plugin/bootstrap/autoload.php';
+        $app = require __DIR__ . '/../plugin/bootstrap/app.php';
+        $kernel = $app->make('Illuminate\Contracts\Console\Kernel');
+        
+        $kernel->call('optimize:clear');
+    } catch (Exception $e) {
+        echo "<pre>Exception: ", $e->getMessage(), "</pre>\n";
+    }
 }